Updated openstack/openstack Project: openstack-infra/project-config  f60fcf91488087b579e0e1640f7c88be757ebd67
Keystone eventlet gate only in keystone

We need to test keystone running under eventlet until we can get rid
of it, but that doesn't mean that every project needs to test
keystone running under eventlet since a change in other projects
can't break keystone. Only keystone needs to run this job.

Eventually this will allow changing the gate to run the
shared jobs in a consistent configuration and it will be easier
to debug gate issues.

I copy-pasted the job template for
'{pipeline}-tempest-dsvm-nova-wsgi-full{branch-designator}' and
switched USE_MOD_WSGI to KEYSTONE instead.

The job should be gating on all branches. Keystone has had eventlet
support in all supported branches. We've been gating on it in the
postgres job since that was switched to use eventlet.

Change-Id: I4599c9864d5c5615ee8e380d05a19cab7861f17c