Updated openstack/openstack Project: openstack-infra/project-config  68f326eef3639543892a5e82694ee37504bebf1c
Move neutron-large-ops jobs to neutron's experimental queue

The existing neutron-large-ops spawns multiple VMs at the same time,
which results in GET requests being flooded to the server.

However, this is done without vNICs, without exercising the Neutron
backends, and without any pre-existing resource load. For this reason,
this job does not quite pounds Neutron hard enough, and the chance of
breaking it with Neutron patches ends up being practically nil. Chances
are that a potential regression is caught by some other job, like Rally.

This change proposes to demote the job to the experimental queue
for Neutron, ideally to be replaced by a job that can more effectively
put strain on the server, and under different circumnstances (for
instance with prepolated load). The job should be fine at catching
regressions induced by how a client calls Neutron, and for this
reason, it's left on the large-ops for Nova and the other projects.

For now, let's free up a slot for something else.

Change-Id: I5f31ab0e29cafc32b7539508413c33238d7af9fa