Updated openstack/openstack Project: openstack/nova  82f2dfe2285ca6bf73f5a706bcb03fe42500f7ff
Don't load deleted instances

When the HostManager starts up, it loads the instance information for
all of the compute nodes. The call it uses, though, includes the records
for deleted instances, which are not needed, and which can cause
performance issues for large deployments. This patch adds a filter on
deleted records.

Closes-bug: 1524421

Change-Id: I404b6baa5504316f447d9c504132183198ff689e