Updated openstack/openstack Project: openstack/nova  845834d810384ef54d6f56ee2d66833bb4c49d5c
Drop JSON for stats in virt API

The ironic virt driver passes the value for stats to the
resource tracker as a JSON encoded string. The resource
tracker checks for this and immediately decodes it.
The JSON is not used anymore since compute node object was
added.

This patch removes that unnecessary encoding and decoding.

blueprint mitaka-objects

Change-Id: I03a8359c690ce00f0b3099155ada7aff5ad5c78a