Updated openstack/openstack Project: openstack/nova  fff885207977d9c1382157bb5bd99182427a16ee
libvirt: Drop JSON for supported_instances

The virt drivers pass the value for supported_instances
to the resource tracker as a JSON encoded string. This
gets decoded as soon as it is copied to the compute node
object. The JSON is not used anymore since compute node
object was added so we will remove the encoding for all
virt drivers.

This patch removes that unnecessary encoding from the libvirt
driver.

blueprint mitaka-objects

Change-Id: Ib7a65da54ac5e08f1f658850465c143e44d5a85c