Updated openstack/openstack Project: openstack/nova  5fffcd0f467e10471222b13d317ebb641dd20284
ironic: Drop JSON for supported_instances

The virt drivers pass the value for supported_instances
to the resource tracker as a JSON encoded string. This
gets decoded as soon as it is copied to the compute node
object. The JSON is not used anymore since compute node
object was added so we will remove the encoding for all
virt drivers.

This patch removes that unnecessary encoding from the
ironic driver.

blueprint mitaka-objects

Change-Id: I2abd20b3beba190896325bb82a6fa94f5aea47e1