Updated openstack/openstack Project: openstack/nova  6e18eeb7a5459963c743dc2550aaec4c11fd551a
Remove XML description from extension concept

XML API interface has been already removed from Nova completely,
and its description of the extension concept becomes meaningless now.
So this patch removes it.

A small amount of the JSON description is left. The details are
trimmed down, as we're actually trying to get people away from
extensions as a concept.

A big warning to that effect is added into the document.

Partially implements blueprint complete-todo-in-api-concept-doc

Change-Id: Icf8bf87f38733ba03cf5ce1225594b0b282fde56