Updated openstack/openstack Project: openstack/tripleo-image-elements  93d18be0f965fed963795d46b991cc0c6115ac16
os-net-config: add configure_safe_defaults

This patch adds a new configure_safe_defaults function that
is called on EXIT. If the exit code is non-zero the function
executes code to configure DHCP on all active interfaces.

Since DHCP is used to initially bootstrap all TripleO instances
via the provisioning network this should always be a safe default
and allow communication with Heat to persist if an invalid
(breaking) network configuration change is pushed out. This would
facilitate immediate Heat notification of the failure and subsequent
Heat updates to potentially help correct the invalid network
configuration.

Change-Id: Ibccba1ee77aa13c85a78a75d2e6cd179f87d7d16
Depends-on: Ibe0e32bc09979bc68b92a722b2bfa383e77502a9