Updated openstack/openstack Project: openstack/neutron  495f8c606e923a758645692045cddf070b6a7a7f
Use a joined relationship for AZ info on routers

This patch adds a joined relationship from the routers
to the agent table to prevent excessive DB lookups of the
agent availability zones during router listings.

This fix is the same as the one for networks here:
I908ceb1a68e0eed7c304e3ff82279ad6fa406167

A follow-up patch will add an AZ table so this doesn't
have to join directly to the agents table.

Change-Id: I808f8ff0c581478b2ea1144e297d83af599a32f2
Closes-Bug: #1526644