Updated openstack/openstack Project: openstack/nova  d7f2aee300e8e4fe36ee4569fc2a36401ad51a62
Implement GlanceEndpoint object

This implements a GlanceEndpoint object to allow us to transition from
internally passing around unstructured tupples to something which
looks more like a url. This object implements properties and an
as_tupple method to let it be largely used as is in existing code.

There are no unit tests for this new object because it's transitional
scaffolding to create an easier to review patch series and removed in
If963b5c7abc132afab259783b13c6d771d2e2897

Change-Id: If53ea16ea71b3ef9a3d81cf1f336ad6b33490e69