Updated openstack/openstack Project: openstack/nova  7b6f264aa320024327ce5d91bdf8a1916f11e676
always create clients with GlanceEndpoint

Instead of building a url on the fly to hand to glanceclient.Client,
change all the calling code to use GlanceEndpoint instead. This is
done through a new _glanceclient_from_endpoint so we can piecemeal
route around it.

Unit test fixed where needed, now that we're calling a different
helper function.

drop _create_glance_client function

This is now no longer used. This drops the function and updates all
the unit tests to use _glanceclient_from_endpoint instead.

The unit tests for special casing the non keystone case is dropped, as
sending the additional headers should be fine in the noauth case.

Change-Id: Idf6050cb9d85b764a1e0a712528a91d1d3d025d1