Updated openstack/openstack Project: openstack/nova  63c18128682fc5e07cb415f1b9f6fda628ee442f
retool xen glance plugin to work with urls

This makes internal changes to the xen glance plugins so that they are
internally passing information about the glance servers by urls. It
does not change the surface of the plugin (that will come later), so
at a top level we are still thinking of glance as a host, port, and
use_ssl. However that's immediately converted to an endpoint url which
is then used for all the rest of the effort.

Change-Id: Ie1dce099846c725ff1ad7bb32f74aa6ae701ee9a