Updated openstack/openstack Project: openstack-infra/zuul  13cb40c3a076ba78edc5e35729c6bd7ef38ff9bb
Add exception handler to updateBuildDescriptions

Adding a exception handler to updateBuildDescriptions to
handle exceptions from launcher (Gearman).

This addresses the corner case seen in:
https://storyboard.openstack.org/#!/story/2000445

In the above mentioned corner case, Gearman throws a communication
error exception. Since updateBuildDescriptions is called after the
build status has been reported to Gerrit, we want to make sure that
we return back to caller so that the build is properly marked as
reported in reportItem to avoid an infinite loop.

Change-Id: I166d4a2e1d45e97a9ed0b6addb7270e1bf92d6f7