Updated openstack/openstack Project: openstack/tooz  da0aa69a7ce34387ad2b842af588481f973bb174
Add basic file content schema validation

Using jsonschema we can validate (at a basic level)
that the files created and the contents written were
not altered to be of a different format so add some
basic format validation to help in detecting format
issues/changes.

This also avoids having to use type errors to try to
detect this same kind of tampering/data format change.

Change-Id: I80b4e560f24162e079cf4adab06d35ec75f4b70c