Updated openstack/openstack Project: openstack/ironic  35e9e24e93f8ce111abfea99687116ebbadb7e95
Devstack: Remove QEMU hook at ./unstack

Remove the libvirt QEMU hook to rotate the logs when ./unstack is
invoked. Otherwise it will always rotate the logs even if the
IRONIC_VM_LOG_ROTATE is set to False in a subsequently ran.

This commit also makes the path to the hook configurable.

Change-Id: Ic0c32e7809f4f0632b6a0bd1e12213aec2a50f84