Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  1e1515001359e47307abea3ef93e942d3243a0ae
Add support for matrices' subset in multijobs.

Let a multijob phase specify a subset of a matrix job via a
Groovy filter, as the Parameterized Trigger plug-in does.
This is from a builder, not a publisher.

A project can define a filter via 'restrict-matrix-project'.

Two plug-ins are involved in this: Parameterized Trigger and
Multijob.

Change-Id: Ifa5f91ed080c4eb43c31a593b70b5e1e5169e77e
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>