Updated openstack/openstack Project: openstack/oslo.config  f891a606b04d9dcad21ad2fa7e3d31dee49c0192
Clarify option types documentation

The documentation for "Option Types" was not clear. There was a list
of types, but it was unclear what it was referencing, the types
(oslo_config.types.*) or the options (oslo_config.cfg.*Opt).

To make clarify, a new table shows what the type is for the provided
convenience option classes.

Change-Id: I4bbc2e29ba5576e9605e1ecaf0dc1ac987580b2f