Updated openstack/openstack Project: openstack/manila  b673bc3206776efda996c8b81bd6a9c6706fb6e7
glusterfs_native: fix parsing of the dynamic-auth option

The full name of the 'dynamic-auth' GlusterFS option
is 'server.dynamic-auth'. When we started to make use
of it in https://review.openstack.org/229409, we just
used its short form, 'dynamic-auth'. That actually
*should* work but due to some weakness in the GlusterFS
instrumentation code, it doesn't.

Fix the situation for now by using the full option
name. We also take the opportunity to remove some
redundant checks from the respective option handling
code.

Change-Id: I6f30060cec38733afa1211cb8e7bc57b66dead34
Closes-Bug: #1506577