Updated openstack/openstack Project: openstack/nova  3b12c70689435acf04c95873a54aa6a22b7ce9a2
Use just ids in all request templates for flavors/images

While the API allows for strings which happen to end in an ID for
flavorRef/imageRef, the common path is really just the ID itself, and
we should have the samples do this. The use of full urls in our
samples makes it seem like this is more meaningful than it really is,
and that you could pass a url to a remote system, for instance, which
won't work at all.

Change-Id: I669619b627f954232712d54498b3744353d25451