Updated openstack/openstack Project: openstack/ironic  d7c41ebf9bc07f253fe08f77fde62a5a7ffef776
Fix "No closing quotation" error when building with tox

Apparently # is treated as a comment inside the string and as
consequence the rest of the command gets stripped. This patch drops
those comments from the bashate command.

This patch also moves the bashate command to its own script under
tools/.

Change-Id: Ic44bb64419ca1465f69c91fd579e8c84d69f84b8
Closes-Bug: #1527562