Updated openstack/openstack Project: openstack/nova  c5c99fe968cfff9103c3376c69f61c52c06a1df4
remove nova-cert from most functional tests

nova-cert is only needed for 1 API resource (os-certificates) and the
ec2 api. We don't need to run it on all the other functional
tests. Getting rid of a service start / stop on all of them shaves a
little time off each test, and means noise in logs.

Change-Id: I175b8913fe5d19c8b552c60c793565c042c6b874