Updated openstack/openstack Project: openstack/sahara  b5453c286439fac68d08d0df1260171c845cadc0
Don't configure hadoop.tmp.dir in Spark plugin

During job execution with Swift spark-submit writes some data in
hadoop.tmp.dir. But for executor user there is no permissions for
writing in current directory. So, it's suggested to use default value
for this config options as it's done for all other plugins.

Change-Id: I0e2ec932cfd4fc49cd0f20098badb15c18769e20
Closes-bug: 1524997