Updated openstack/openstack Project: openstack/heat-cfntools  5684248e19e7fcac615b987771a6595eb6a0b29b
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then, run
py27. This patch puts py34 first on the tox.ini list of envs to avoid this
problem to happen.

Change-Id: Id0d2de40081a59d5819d2e6c52c98135f26808b5
Closes-bug: #1489059