Updated openstack/openstack Project: openstack/heat-translator  4ecee19a8aa0dd97aa11aff40ddc34f22304e8a4
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then, run
py27. This patch puts py34 first on the tox.ini list of envs to avoid this
problem to happen.

Change-Id: I5c838c6f8e9288372ee83cfb993df4e077e75819
Closes-bug: #1489059