Updated openstack/openstack Project: openstack/openstack-doc-tools  d1381c57df6865da852c4a00a700fca36d8c673c
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then, run
py27. This patch puts py34 first on the tox.ini list of envs to avoid this
problem to happen.

Change-Id: I966264e04fbe9441d90c48645df5e6216ca0f4d5
Closes-bug: #1489059