Updated openstack/openstack Project: openstack/pycadf  043b209b46a2e4237eb05e6e6a44767b4a8606a6
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then, run
py27. This patch puts py34 first on the tox.ini list of envs to avoid this
problem to happen.

Change-Id: I461e4f0d931a9bbc37670dc2cd70d94f86acfb24
Closes-bug: #1489059