Updated openstack/openstack Project: openstack/python-designateclient  90a01cb2c53e934dec18783f8708497d8ec96653
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then, run
py27. This patch puts py34 first on the tox.ini list of envs to avoid this
problem to happen.

Change-Id: If022a76b2ee9a290033abc544c5fa56f1ea1e0ae
Closes-bug: #1489059