Updated openstack/openstack Project: openstack/neutron  252596cbedad2dac53ff9395d20a7128a422d3f1
Use the constant HOST_ID instead of 'binding:host_id'

This patch makes use of the constant defined in the extension.
In addition to this having value of debing defined in one place it
also enables the caller to understand that the portbindings
extension is required.

Note: the constant is not used in the API tests. This has import
issues so it is not relevant.

TrivialFix

Change-Id: I7bfe2528dbbd8017ddbdcf949dbb6264ce1eb5d8