Updated openstack/openstack Project: openstack/tripleo-heat-templates  92c994deb8a3719cf7853f46b0aa0b52489fc368
Add SoftwareConfigTransport for switching transports

This change adds a SoftwareConfigTransport parameter to role templates
so that the transport can be changed via a parameter_defaults entry.

This change will have no effect on an existing overcloud as the current
default POLL_SERVER_CFN is now explicit in the parameter default.

Change-Id: I5c2a2d2170714093c5757282cba12ac65f8738a4