Updated openstack/openstack Project: openstack/neutron-lbaas  961ada07435d94c94c42b48fa989b32bc8374078
Fix db refresh race condition bug

There were a few test failures that exposed a race condition when an entity was
retrieved from the db twice.  The second retrieval was on a session refresh.
If a delete of that entity occurred right before the refresh, the refresh would
fail because that entity did not exist anymore.

The solution is to simply not do two reads.  Combined with the patch
in neutron that this depends on, this will fix the reason the refresh
existed and the race condition.

Closes-Bug: #1504465
Change-Id: If160d65df3448e2790fc869d7293cc750f6775b7