Updated openstack/openstack Project: openstack/python-glanceclient  9a9f334920105319d8745918cb32b0c105ce5463
Disable suggestion of v1 help for v2 commands

Currently the client suggests 'Run `glance --os-image-api-version 1
help` for v1 help' at the end of every help message. This is could
be confusing for a v2 only command.

Therefore this patch disables it if the command does not exist in v1,
while keeping the message on the 'glance help' results.

Change-Id: I967e9ba35afb8dc40524bd1d13284e684b435f81
Closes-Bug: 1520602