Updated openstack/openstack Project: openstack/manila  e5a6f05b03f21b3770bdd63e0e2ed042482f4ec4
Capability lists in Manila scheduler

Some share drivers would like to report pool capabilities such as
'compression': [True, False] to denote a pool which can simultaneously
handle shares with compression enabled and others with compression
disabled. This commit enhances the CapabilitiesFilter to handle this
use case.  An extra spec value will be compared to a capability list
and will be passed by the filter if it matches anything in the list.

Implements: blueprint support-lists-in-capabilities-filter
Change-Id: I50963a7cde54c10632ecf5caba2ee436073da8f8