Updated openstack/openstack Project: openstack/taskflow  25a8e4b13253cd13d1c9590c8e5415c2a3204e00
Fix currently broken and inactive mysql tests

Commit 868dd8485456ce78a396b687ec7744dc365f06be added a
migration script that caused there to be two heads (starting
points) of migrations which would cause alembic to be unable
to do the initial schema creation (and upgrade).

So this fixes that by correctly making one head only and
not skipping tests for a database if it can not be
upgraded for whatever reason (which was silencing the
test failures).

Closes-bug: #1528683

Change-Id: Id571072eec1dc3b6cbb9e868854f6db0b271e5f8