Updated openstack/openstack Project: openstack-infra/shade  c10566cc0ab0a9415d650a93941deb9b62ecdaf7
Fix shade tests with OCC 1.13.0

The latest os-client-config, 1.13.0, breaks the shade tests as the
client instantiations are slightly different. Since I'm not sure we
can modify our tests to use against different OCC versions (there
doesn't appear to be a version attribute anywhere), I upped the
minimum requirements.txt value as well as fixed the tests for the
latest version.

Change-Id: I18292ca39789c41caea74901754d66d3301bb420