Updated openstack/openstack Project: openstack/glance  8d483acdbcc956b1b411684bdf5854a9b8fe2eb4
Add sign-the-hash deprecation warning

Per discussion on the mailing list [1] and the related nova specification [2]
it has been decided that the signature should be of the image data directly,
rather than of the glance MD5 "checksum" hash of the image data.

This patch adds TODO statements to remove the 'sign-the-hash' approach
after the Mitaka development cycle, as well as a deprecation warning if
the properties exist for the 'sign-the-hash' approach.

[1] http://bit.ly/1Q0M0C7
[2] https://review.openstack.org/#/c/188874/19

Partial-Bug: #1516031
Change-Id: Ifaae6e34491b382ffffe5b4e4764764d592d1f53