Updated openstack/openstack Project: openstack/neutron  ce703f7b91e55ff2c6a61bf842f28d7d41cf190a
Force L3 agent to resync router it could not configure

If the L3 agent fails to configure a router, commit:
4957b5b43521a61873a041fe3e8989ed399903d9 changed it so
that instead of performing an expensive full sync, only that
router is reconfigured. However, it tries to reconfigure the
cached router. This is a change of behavior from the fullsync
days. The retry is more likely to succeed if the
router is retrieved from the server, instead of using
the locally cached version, in case the user or operator
fixed bad input, or if the router was retrieved in a bad
state due to a server-side race condition.

Note that this is only relevant to full syncs, as those retrieve
routers from the server and queue updates with the router object.
Incremental updates queue up updates without router objects,
so if one of those fails it would always be resynced on a
second attempt.

Related-Bug: #1494682
Change-Id: Id0565e11b3023a639589f2734488029f194e2f9d