Updated openstack/openstack Project: openstack-infra/project-config  e38327227eff6bc6fb7ca0e2d2da377806969928
Replace bashate/pep8 in project-config by linters

Use a single linting target for changes and run that instead of the
individual bashate and pep8 targets.

Fix name of pep8 job for project-config-example.

Create new gate-{name}-linters job-template for this.

The pep8 job was run also on project-config-examples, create a new job
for that repo. It now runs without the filters since with the amount
of changes to the repo, we do not need to optimize anything right now.

Change-Id: I2afa7931bf3e18798b95a8af33ccf5ef761c4ad5