Updated openstack/openstack Project: openstack/oslo.vmware  342b0913b4239d08e35c8f64a0589d5ce8f10509
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then, run
py27. This patch puts py34 first on the tox.ini list of envs to avoid this
problem to happen.

Change-Id: I9dd91b4b58e84c9f73b1e662aa24a779629a8a37
Closes-bug: #1489059