Updated openstack/openstack Project: openstack-infra/project-config  184e6efca5051cbfe18f30a690ecac56ffd5ece5
Build a Debian simple-init images

In order to to speed CI build times for bifrost, we will
attempt to pre-build the debian guest image that helps us
identify if simple-init is in working order as compared to
bifrost, since the simple-init element which users are
dependent upon resides in the diskimage-builder repository.

As a result of the same job being utilized, the job has been
renamed appropriately.

Additionally, the file upload configuration appeared to be
working incorrectly and this commit updates it in an attempt
to correct the issue.

Change-Id: Ifde8717740f324cb3abe1266efca42bd24d634d3