Updated openstack/openstack Project: openstack-infra/project-config  a2b6ff64d41628e9cb62502de274219a8010f3ed
Make install-distro-packages.sh more robust

Introduce a retry loop in install-distro-packages.sh which will
attempt index updates and package installation up to 3 times if the
bindep list is unsatisfied, so as to mitigate intermittent network
failures. Add a clear error message after the last try indicating
the list of packages which were not installed.

Change-Id: I595aedbc57623ee1b30bd5880abf699f95e9eb63