Updated openstack/openstack Project: openstack/tempest  9db196c625b928706aad72496fc4c998c8b0b9c9
Stop testing rebuild for ironic

Currently our tests run too long and we experience constant timeouts.
Rebuild is a relatively rarely used feature, so we plan on creating
separate jobs for it. Remove rebuild testing completely with the goal
to reintroduce it as soon as we switch to an in-tree tempest plugin.

Change-Id: I16d27b6bd900186b84df7045a6e4f66257fefac4