Updated openstack/openstack Project: openstack/neutron  bf9e4133eb787fdcc43692a98aea63f15d37c7a1
Add address scope to floating IPs in RPC response to L3 agent

The L3 agent needs to know the address scope of the fixed ip of each
floating ip because floating ips are a way to cross scope boundaries.
Without the scope information, there could be ambiguity and no way to
know which scope to send it to.

[1] https://review.openstack.org/#/c/189741/

Change-Id: Id9f8c12954a6efbf4d9b99c011652eefbe5f5145
Partially-Implements: blueprint address-scopes