Updated openstack/openstack Project: openstack/neutron  89f96e627fc1e402ca456d60127ccde5bd534415
Ensure agent binding modules are loaded

This patch ensures that models_v2 and the router db module
import the agent binding modules so their models are defined
in the sqlalchemy metadata. Otherwise, using an agentless
core/L3 plugin may result in a failure to initialize if nothing
imported the agent module.

Change-Id: I77240462b618d6e95c7c5862441d6d179e371f4f
Closes-Bug: #1529022