Updated openstack/openstack Project: openstack/manila  e382d717112a5a886cf7a8dca0f1308b122e2c72
Pop off user/tenant kwargs in RequestContext init

The RequestContextSerializer uses the to_dict and from_dict
methods and to_dict is storing user/tenant fields which end
up unaccounted for in the RequestContext constructor kwargs
which means we log a warning every time we hydrate the
context - which is a lot (~89K times in 2 days of CI runs).

This change simply pops the user/tenant from kwargs before
logging the warning and uses them in case user_id or
project_id are None for some reason, which is similar to what
Nova's RequestContext does.

Change-Id: I57b4854d60d5758a67df5edcdcddeeb0a4ad886d
Closes-Bug: #1528348