Updated openstack/openstack Project: openstack-infra/shade  a39fca8e53e7de0c4ffdeafce40902e0e6fc7073
Fix unittest stack status

If you run unittests with a clouds.yaml available that has caching
enbaled, the heat test fails because the fake has an invalid value for
stack_status (it'll always be a string) and also the pending function
looks for the wrong name.

We should isolate unittests from calling context clouds.yaml better, but
this is a bug anyway.

Change-Id: I791df68163b9ce850df6c74eacadd91ff76348ed