Updated openstack/openstack Project: openstack/neutron  550ab90c1b0424ffa0d37a3d18b7c074a01e22cc
Add tests that constrain db query count

This patch adds unit tests to ML2 and L3 that ensure that the
number of DB calls during list operations for ports, networks,
subnets, routers, and floating IPs remains constant regardless
of the number of ports.

These will prevent changes from slipping in that result in
a separate DB query for each object in a list operation
(for changes to the extensions used by ML2 and the DVR plugin).

Related-Bug: #1525295
Related-Bug: #1513782
Related-Bug: #1525423
Related-Bug: #1525740
Related-Bug: #1526644

Change-Id: I1958fc7c318bbf73238a3ad5be133fa7800c8290