Updated openstack/openstack Project: openstack-infra/subunit2sql  c37eb3d97d8d1adbc3a2e66659206c2e0152ac67
Stop using implict joins

This commit stops the usage of implict joins everywhere throughout
the subunit2sql codebase. When we dropped the use of foreign keys
in the db schema with the big uuid -> int migration fks were left
in the sqlalchemy models because implict joins were in use in the
db api. This commit fixes this so all the joins are explict and the
foreign keys can be removed from the db models.

Change-Id: Ia20961a154758fbfc74aaefe2103fc35c5b6db27