Updated openstack/openstack Project: openstack/keystone  255b317ffd22f09a9b94b8c169cd5c05af133ee6
Correct the class name of the V9 LDAP role driver

In the recent change to create a V9 role driver
(https://review.openstack.org/#/c/247805/) the renaming to V9
was missed for the LDAP driver (the SQL driver was correct).

This patch fixes the LDAP driver class name.

Partially Implements: blueprint implied-roles

Change-Id: I49efeedaa5c163c4434dacc3c3467497d9fffca2